Big reorganization, especially CSS
adopt project structure technique I learned at The Company That Shall Not Be Named
