+++ frontend/src/components/Details/Details.js
@@ -1,5 +1,5 @@
 import React from 'react';
 
-import { container } from '../stylesheets/details.css';
+import { container } from './style.css';
 
 export default ({ children }) => <div className={container}>{children}</div>;
