+++ frontend/src/components/Details/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "./Details.js",
+  "private": true
+}
