+++ frontend/src/components/Details/style.css
@@ -1,7 +1,7 @@
-@value margin from "./measurements.css";
+@value margin from stylesheets/measurements;
 
 .container {
-  composes: roboto from './shared.css';
+  composes: roboto from '../../stylesheets/shared';
   display: grid;
   grid-template-columns: 20% 1fr;
   grid-template-rows: auto;
