+++ frontend/src/components/Header/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "./Header.js",
+  "private": true
+}
