+++ frontend/src/components/Loading/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "./Loading.js",
+  "private": true
+}
