+++ frontend/src/containers/App/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "./App.js",
+  "private": true
+}
