+++ frontend/src/containers/NotFound/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "./NotFound.js",
+  "private": true
+}
