--- frontend/src/containers/__tests__/Home-test.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-
-import Home from '../Home';
-
-describe('home component', () => {
-  it('matches snapshot', () => {
-    expect(shallow(<Home />)).toMatchSnapshot();
-  });
-});
