--- frontend/src/stylesheets/link.css
@@ -1,19 +1,0 @@
-@value black, blue, dark-blue from colors;
-
-.blacklink {
-  color: black;
-  text-decoration: none;
-
-  &:hover {
-    text-decoration: underline;
-  }
-}
-
-.link {
-  composes: blacklink;
-  color: blue;
-
-  &:visited {
-    color: dark-blue;
-  }
-}
