+++ frontend/src/stylesheets/shared.css
@@ -1,4 +1,4 @@
-@value black from './colors.css';
+@value black, blue, dark-blue from colors;
 
 .center {
   display: flex;
@@ -27,5 +27,23 @@
   &:hover:not(:disabled) {
     transform: scale(1.02);
     box-shadow: 2px 2px 1px 1px rgba(0, 0, 0, 0.1);
+  }
+}
+
+.blacklink {
+  color: black;
+  text-decoration: none;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.link {
+  composes: blacklink;
+  color: blue;
+
+  &:visited {
+    color: dark-blue;
   }
 }
